Print multiple ships per row when possible
For a quick build with multiple ships that fit on a single row, do it.

For a single ship with a generic pilot, repeat it as many times as fits
on one row. We assume these are low threat level and there is a chance
a player might want multiple copies. If a player only wants 1 ship, they
will need to obscure the others (with the other quick build cards they
have).